refactor/ Create utils folder for better UI architecture
- Create constants file to have the array for bookshelves

- Map through the bookshelves to create the view instead of duplicating the code

- Move the API file to the utils folder
