+++ starter/src/pages/main-page.js
@@ -2,6 +2,7 @@ import React from 'react'
 import Bookshelf from '../components/bookshelf';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
+import { BOOKSHELVES } from '../utils/constants';
 
 function MainPage({ books, onChangeBookshelf }) {
     return (
@@ -11,18 +12,12 @@ function MainPage({ books, onChangeBookshelf }) {
             </div>
             <div className="list-books-content">
                 <div>
-                    <Bookshelf
-                        books={books.filter((book) => book.shelf === "currentlyReading")}
-                        bookshelfTitle={"Currently Reading"}
-                        onChangeBookshelf={onChangeBookshelf} />
-                    <Bookshelf
-                        books={books.filter((book) => book.shelf === "wantToRead")}
-                        bookshelfTitle={"Want To Read"}
-                        onChangeBookshelf={onChangeBookshelf} />
-                    <Bookshelf
-                        books={books.filter((book) => book.shelf === "read")}
-                        bookshelfTitle={"Read"}
-                        onChangeBookshelf={onChangeBookshelf} />
+                    {BOOKSHELVES.map((bookshelf) => (
+                        <Bookshelf key={bookshelf.value}
+                            books={books.filter((book) => book.shelf === bookshelf.value)}
+                            bookshelfTitle={bookshelf.label}
+                            onChangeBookshelf={onChangeBookshelf} />
+                    ))}
                 </div>
             </div>
             <div className="open-search">
