+++ starter/src/utils/constants.js
@@ -1,0 +1,14 @@
+export const BOOKSHELVES = [
+    {
+        label: "Currently Reading",
+        value: "currentlyReading"
+    },
+    {
+        label: "Want To Read",
+        value: "wantToRead"
+    },
+    {
+        label: "Read",
+        value: "read"
+    },
+]
